tried to reduce slack api service, in size to improve testability
